+++ packages/block-editor/src/components/inner-blocks/index.native.js
@@ -50,8 +50,9 @@ function UncontrolledInnerBlocks( props ) {
 		horizontalAlignment,
 	} = props;
 
-	const block = useSelect( ( select ) =>
-		select( 'core/block-editor' ).getBlock( clientId )
+	const block = useSelect(
+		( select ) => select( 'core/block-editor' ).getBlock( clientId ),
+		[ clientId ]
 	) || { innerBlocks: [] };
 
 	useNestedSettingsUpdate( clientId, allowedBlocks, templateLock );