+++ packages/block-library/src/image/image-editor.js
@@ -12,26 +12,30 @@ import classnames from 'classnames';
 import { BlockControls } from '@wordpress/block-editor';
 import { useState } from '@wordpress/element';
 import {
+	Icon,
+	search,
 	rotateRight as rotateRightIcon,
 	aspectRatio as aspectRatioIcon,
 } from '@wordpress/icons';
 import {
 	ToolbarGroup,
 	ToolbarButton,
-	__experimentalToolbarItem as ToolbarItem,
 	Spinner,
 	RangeControl,
 	DropdownMenu,
 	MenuGroup,
 	MenuItem,
 } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
+import { __, sprintf } from '@wordpress/i18n';
 import { useDispatch } from '@wordpress/data';
 import apiFetch from '@wordpress/api-fetch';
 
 const MIN_ZOOM = 100;
 const MAX_ZOOM = 300;
-const POPOVER_PROPS = { position: 'bottom right' };
+const POPOVER_PROPS = {
+	position: 'bottom right',
+	isAlternate: true,
+};
 
 function AspectGroup( { aspectRatios, isDisabled, label, onClick } ) {
 	return (
@@ -51,13 +55,13 @@ function AspectGroup( { aspectRatios, isDisabled, label, onClick } ) {
 	);
 }
 
-function AspectMenu( { isDisabled, onClick, toggleProps } ) {
+function AspectMenu( { isDisabled, onClick } ) {
 	return (
 		<DropdownMenu
 			icon={ aspectRatioIcon }
 			label={ __( 'Aspect Ratio' ) }
 			popoverProps={ POPOVER_PROPS }
-			toggleProps={ toggleProps }
+			className="wp-block-image__aspect-ratio"
 		>
 			{ ( { onClose } ) => (
 				<>
@@ -164,14 +168,20 @@ export default function ImageEditor( {
 	function apply() {
 		setIsProgress( true );
 
-		const attrs = crop;
+		let attrs = {};
+
+		// The crop script may return some very small, sub-pixel values when the image was not cropped.
+		// Crop only when the new size has changed by more than 0.1%.
+		if ( crop.width < 99.9 || crop.height < 99.9 ) {
+			attrs = crop;
+		}
 
 		if ( rotation > 0 ) {
 			attrs.rotation = rotation;
 		}
 
 		apiFetch( {
-			path: `__experimental/image-editor/${ id }/apply`,
+			path: `wp/v2/media/${ id }/edit`,
 			method: 'POST',
 			data: attrs,
 		} )
@@ -182,10 +192,12 @@ export default function ImageEditor( {
 					height: height && width ? width / aspect : undefined,
 				} );
 			} )
-			.catch( () => {
+			.catch( ( error ) => {
 				createErrorNotice(
-					__(
-						'Unable to perform the image modification. Please check your media storage.'
+					sprintf(
+						/* translators: 1. Error message */
+						__( 'Could not edit image. %s' ),
+						error.message
 					),
 					{
 						id: 'image-editing-error',
@@ -259,8 +271,26 @@ export default function ImageEditor( {
 
 	return (
 		<>
+			{ ! inProgress && (
+				<div
+					className="wp-block-image__zoom-control"
+					aria-label={ __( 'Zoom' ) }
+				>
+					<Icon icon={ search } />
+					<RangeControl
+						min={ MIN_ZOOM }
+						max={ MAX_ZOOM }
+						value={ Math.round( zoom ) }
+						onChange={ setZoom }
+					/>
+					<AspectMenu
+						isDisabled={ inProgress }
+						onClick={ setAspect }
+					/>
+				</div>
+			) }
 			<div
-				className={ classnames( 'richimage__crop-area', {
+				className={ classnames( 'wp-block-image__crop-area', {
 					'is-applying': inProgress,
 				} ) }
 				style={ {
@@ -277,23 +307,15 @@ export default function ImageEditor( {
 					zoom={ zoom / 100 }
 					aspect={ aspect }
 					onCropChange={ setPosition }
-					onCropComplete={ setCrop }
+					onCropComplete={ ( newCropPercent ) => {
+						setCrop( newCropPercent );
+					} }
 					onZoomChange={ ( newZoom ) => {
 						setZoom( newZoom * 100 );
 					} }
 				/>
 				{ inProgress && <Spinner /> }
 			</div>
-			{ ! inProgress && (
-				<RangeControl
-					className="richimage__zoom-control"
-					label={ __( 'Zoom' ) }
-					min={ MIN_ZOOM }
-					max={ MAX_ZOOM }
-					value={ Math.round( zoom ) }
-					onChange={ setZoom }
-				/>
-			) }
 			<BlockControls>
 				<ToolbarGroup>
 					<ToolbarButton
@@ -302,17 +324,6 @@ export default function ImageEditor( {
 						onClick={ rotate }
 						disabled={ inProgress }
 					/>
-				</ToolbarGroup>
-				<ToolbarGroup>
-					<ToolbarItem>
-						{ ( toggleProps ) => (
-							<AspectMenu
-								toggleProps={ toggleProps }
-								isDisabled={ inProgress }
-								onClick={ setAspect }
-							/>
-						) }
-					</ToolbarItem>
 				</ToolbarGroup>
 				<ToolbarGroup>
 					<ToolbarButton onClick={ apply } disabled={ inProgress }>