+++ packages/components/src/utils/hooks/test/use-controlled-state.js
@@ -7,14 +7,16 @@ import { noop } from 'lodash';
 /**
  * Internal dependencies
  */
-import { useControlledState } from '../hooks';
+import useControlledState from '../use-controlled-state';
 
 describe( 'hooks', () => {
 	const getInput = () => screen.getByTestId( 'input' );
 
 	describe( 'useControlledState', () => {
-		const Example = ( { value: valueProp, onChange = noop } ) => {
-			const [ state, setState ] = useControlledState( valueProp );
+		const Example = ( { initial, value: valueProp, onChange = noop } ) => {
+			const [ state, setState ] = useControlledState( valueProp, {
+				initial,
+			} );
 
 			const handleOnChange = ( event ) => {
 				const nextValue = event.target.value;
@@ -44,7 +46,7 @@ describe( 'hooks', () => {
 
 		it( 'should update rendered value onChange', () => {
 			const spy = jest.fn();
-			render( <Example value="Hello" onChange={ spy } /> );
+			render( <Example onChange={ spy } /> );
 
 			const input = getInput();
 
@@ -78,9 +80,7 @@ describe( 'hooks', () => {
 			 * from props. The prop being rendered is the state provided by
 			 * useControlledState, rather than the value prop directly.
 			 */
-			const { rerender } = render(
-				<Example value="Hello" onChange={ spy } />
-			);
+			const { rerender } = render( <Example onChange={ spy } /> );
 
 			const input = getInput();
 
@@ -102,23 +102,33 @@ describe( 'hooks', () => {
 			expect( spy ).toHaveBeenCalledTimes( 1 );
 		} );
 
-		it( 'should new incoming prop value with new change value', () => {
+		it( 'should render with initial value and be controllable', () => {
 			const spy = jest.fn();
+			// Input starts off as being uncontrolled / self-managed.
 			const { rerender } = render(
-				<Example value="Hello" onChange={ spy } />
+				<Example onChange={ spy } initial="Hello" />
 			);
 			const input = getInput();
 
+			expect( input.value ).toBe( 'Hello' );
+
 			fireEvent.change( input, { target: { value: 'There' } } );
 
+			expect( input.value ).toBe( 'There' );
+			expect( spy ).toHaveBeenCalledWith( 'There' );
+
+			// Input is now controlled
 			rerender( <Example value="New Value" /> );
+			expect( input.value ).toBe( 'New Value' );
 
 			fireEvent.change( input, {
 				target: { value: 'New Change Value' },
 			} );
 
-			expect( input.value ).toBe( 'New Change Value' );
-			expect( spy ).toHaveBeenCalledTimes( 1 );
+			// Since the Input is being controlled by a parent element,
+			// the internal state is no longer used. Instead, the incoming
+			// prop value will be used.
+			expect( input.value ).toBe( 'New Value' );
 		} );
 	} );
 } );