+++ packages/edit-navigation/src/components/navigation-editor/block-editor-area.js
@@ -36,27 +36,34 @@ export default function BlockEditorArea( {
 	menuId,
 	saveBlocks,
 } ) {
-	const { rootBlockId, isNavigationModeActive, hasSelectedBlock } = useSelect(
-		( select ) => {
-			const {
-				isNavigationMode,
-				getBlockSelectionStart,
-				getBlock,
-				getBlocks,
-			} = select( 'core/block-editor' );
+	const {
+		rootBlockId,
+		isNavigationModeActive,
+		isRootBlockSelected,
+		hasSelectedBlock,
+	} = useSelect( ( select ) => {
+		const {
+			isNavigationMode,
+			getBlockSelectionStart,
+			getBlock,
+			getBlocks,
+		} = select( 'core/block-editor' );
 
-			const selectionStartClientId = getBlockSelectionStart();
+		const selectionStartClientId = getBlockSelectionStart();
+		const rootClientId = getBlocks()[ 0 ]?.clientId;
 
-			return {
-				rootBlockId: getBlocks()[ 0 ]?.clientId,
-				isNavigationModeActive: isNavigationMode(),
-				hasSelectedBlock:
-					!! selectionStartClientId &&
-					!! getBlock( selectionStartClientId ),
-			};
-		},
-		[]
-	);
+		return {
+			selectionStartClientId,
+			rootBlockId: rootClientId,
+			isNavigationModeActive: isNavigationMode(),
+			isRootBlockSelected:
+				!! selectionStartClientId &&
+				rootClientId === selectionStartClientId,
+			hasSelectedBlock:
+				!! selectionStartClientId &&
+				!! getBlock( selectionStartClientId ),
+		};
+	}, [] );
 
 	const { saveMenu } = useDispatch( 'core' );
 	const menu = useSelect( ( select ) => select( 'core' ).getMenu( menuId ), [
@@ -95,12 +102,15 @@ export default function BlockEditorArea( {
 					className={ classnames(
 						'edit-navigation-editor__block-editor-toolbar',
 						{
-							'is-hidden': isNavigationModeActive,
+							'is-hidden':
+								isNavigationModeActive || isRootBlockSelected,
 						}
 					) }
 					aria-label={ __( 'Block tools' ) }
 				>
-					{ hasSelectedBlock && <BlockToolbar hideDragHandle /> }
+					{ hasSelectedBlock && ! isRootBlockSelected && (
+						<BlockToolbar hideDragHandle />
+					) }
 				</NavigableToolbar>
 				<Popover.Slot name="block-toolbar" />
 				<WritingFlow>